% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ruda.R
\name{inverso_multi}
\alias{inverso_multi}
\title{Function Inverso}
\usage{
inverso_multi(x)
}
\arguments{
\item{x}{An input number to calculate the inverse power}
}
\value{
the inverse power input number
}
\description{
Function Inverso
}
